#128 Aplollo plugin def moved to separate file
